Refactor how selectInputs update
This addresses the problem of the Explore panel plot updating several times
when one input was changed. It was previously reacting to any change to the
project, scenario, query, or subcategory inputs, meaning that in some cases
it was redrawing (or attempting to redraw) up to four times. In this change
each input is chained to the next, and one will only update when the previous
one signals it.